Make a slightly larger TeensyLC program that blinks the on-board LED in Morse code.
Make the LED on the board say "Hello World". This is still a busy loop without any timers,
and it's still a fixed pattern. I eventually want to control the displayed message by commands sent
over USB, and the structure of the code (including some currently-unused tables) is designed to
support that. I also want to make the switch to using interrupts instead of a busy loop.

Replace the copied linker script with one written from scratch. It may look similar,
but that's because there are only so many correct ways to write a linker script.

Add a Makefile.