Tidy up initialization.
First, actually initialize variables in the .data and .bss sections.
This was the whole reason for implementing memset and memcpy.
Why didn't I do this then?

Fix the linker script to get the correct locations of .data and .bss,
so that that initialization works.

Specify our startup test message in the main program, instead of hiding it away in "morse_init".

Change the startup test message so we know we're running the right program.

Remove "morse_init" since it's now empty.