Add functions to create debug messages programmatically.
These functions are all located in "print.S".

I would have liked to implement and use functions from somewhere in the C standard,
but *all* of the most relevant functions in C are a chore to use from assembly.

All I need is the ability to encode numbers as hex strings, and to paste together short strings
into longer ones. The functions I've defined are enough to do that, and I've taken care to give
them an interface that's as painless to use as possible. In doing that, I've violated the AAPCS
by returning values in multiple registers, which is fine since I'm not trying to call these
functions from C.