In 'morse.S', leave at least one unused byte in the transmit buffer at all times.
This is to keep the start or cursor position from being the same as the end position.

If the start position is the same as the end position, that gets interpreted as an empty
buffer (and needs to be interpreted that way since everything is zeroed on startup).
If the cursor position is the same as the end position, that is interpreted as the entire
buffer having been transmitted.